requirements.txt
numpy
torch
torchvision
opencv-python
wandb
pandas
plotly
scipy
h5py
Pillow
tqdm
